revamping urls and navigation
