+++ admin-ui/src/App.tsx
@@ -1,26 +1,26 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { useReactOidc } from "@axa-fr/react-oidc-context";
-import ApplicationRound from "./component/ApplicationRound/ApplicationRound";
+import ApplicationRound from "./component/recurring-reservations/ApplicationRound";
 import PageWrapper from "./component/PageWrapper";
 import "./i18n";
-import Application from "./component/Application/Application";
-import ApplicationDetails from "./component/Application/ApplicationDetails";
-import Recommendation from "./component/ApplicationRound/Recommendation";
-import RecommendationsByApplicant from "./component/ApplicationRound/RecommendationsByApplicant";
-import ApplicationRounds from "./component/ApplicationRound/ApplicationRounds";
-import AllApplicationRounds from "./component/ApplicationRound/AllApplicationRounds";
+import Application from "./component/applications/Application";
+import ApplicationDetails from "./component/applications/ApplicationDetails";
+import Recommendation from "./component/recurring-reservations/Recommendation";
+import RecommendationsByApplicant from "./component/recurring-reservations/RecommendationsByApplicant";
+import ApplicationRounds from "./component/recurring-reservations/ApplicationRounds";
+import AllApplicationRounds from "./component/recurring-reservations/AllApplicationRounds";
 import MainLander from "./component/MainLander";
-import Approval from "./component/ApplicationRound/Approval";
-import Applications from "./component/Application/Applications";
-import Criteria from "./component/ApplicationRound/Criteria";
-import RecommendationsByReservationUnit from "./component/ApplicationRound/RecommendationsByReservationUnit";
-import ApplicationRoundApprovals from "./component/ApplicationRound/ApplicationRoundApprovals";
+import Approval from "./component/decisions/Approval";
+import Applications from "./component/applications/Applications";
+import Criteria from "./component/recurring-reservations/Criteria";
+import RecommendationsByReservationUnit from "./component/recurring-reservations/RecommendationsByReservationUnit";
+import ApplicationRoundApprovals from "./component/decisions/ApplicationRoundApprovals";
 import { publicUrl } from "./common/const";
-import ResolutionReport from "./component/ApplicationRound/ResolutionReport";
-import ReservationsByReservationUnit from "./component/ApplicationRound/ReservationsByReservationUnit";
-import ReservationSummariesByReservationUnit from "./component/ApplicationRound/ReservationSummariesByReservationUnit";
-import ReservationByApplicationEvent from "./component/Application/ReservationByApplicationEvent";
+import ResolutionReport from "./component/recurring-reservations/ResolutionReport";
+import ReservationsByReservationUnit from "./component/recurring-reservations/ReservationsByReservationUnit";
+import ReservationSummariesByReservationUnit from "./component/recurring-reservations/ReservationSummariesByReservationUnit";
+import ReservationByApplicationEvent from "./component/applications/ReservationByApplicationEvent";
 import SpacesList from "./component/Spaces/SpacesList";
 import Units from "./component/Unit/Units";
 import Unit from "./component/Unit/Unit";
