+++ admin-ui/src/component/recurring-reservations/ApplicantBox.tsx
