+++ admin-ui/src/component/recurring-reservations/RecommendedSlot.test.tsx
