+++ admin-ui/src/component/recurring-reservations/RecommendedSlot.tsx
