+++ admin-ui/src/component/recurring-reservations/ResolutionReport.tsx
